// src/shared/types/mock-server-data.ts
import { CiryCoords } from './cities.js';

export type MockServerData = {
  titles: string[],
  descriptions: string[];
  cities: string[];
  coords: CiryCoords[];
  previews: string[];
  photos: string[];
  comforts: string[];
  types: string[];
  names: string[];
  emails: string[];
  avatars: string[];
  passwords: string[];
}